A TUI wait-up script nbb+reagent+ink
This shows each service on a row and shows color highlighting of the
service (running state) in addition to log and cmd checks that are shown
and highlighted based on completion. The checks are defined in a yaml
configuration file.

For a simple example of using nbb with reagent (React) and ink see this
example:
https://github.com/babashka/nbb#reagent