Choose columns with --columns. Some re-org/cleanup
The --columns argument allows selection of which columns to display
(default is all the columns). This required a refactor of the render
generatoring function (renamed to visual-table) to make column headers
and container rows filtered by the column configuration.

Add comma separated parsing function for --columns and use this for
--show-events too (add usage defined defaults for both).

Some minor comment and re-ordering.