Add Code column (exit code). Improve color/styling.
Add an optional Code/:code column. Move Status/:status column to
optional (not default).

Change the cell function to take a full style map rather than just
a color and return a style map from the header, container, and status
color/style functions. Use that to improve the styling of the cells.

Split the styling of the container name from the status and exit-codes.
Make the latter a more muted color of green and red for the exit code..