Bump to 1.3. Keep checks for services with 0 exit code
Some services are one-off services that exit after running to
completion. There are some cases where we want dcmon's finished state to
be able to trigger against those containers after they are completed. In
order to do this we need to keep status check results for exited
containers. In the visual representation, these will be darker green for
successful checks for containers that have exited with a 0 exit code.

Update the example to run for longer before exiting and add some
alternate but commented out finished definitions to be able to test
exiting on successful status checks of exit with 0 containers.