Support containers using tty (e.g. systemd)
For containers that have an attached TTY (Config.Tty: true), the format
of the stream data is different and was breaking the log stream method
we were using. For non-TTY containers, the log stream send chunks with
an 8 byte frame header that specifies the type of chunk buffer (stdout or
stderr) and the chunk length/size. The dockerode modem.demuxStream
method is used to split the chunks into seperate stdout and stderr
streams. For TTY enabled containers, the chunks have no framing and are
just raw log chunk buffers. Using the same method with TTY streams
causes the demuxStream function to effectively wait forever because the
length of the first "frame" tends to be very large (whatever random
4 bytes of ASCII characters are sent at the start of the log).

To fix this we check the inspected container data for the Config.TTY
property and if it's set we just pipe the log stream from the container
to our output stream that we use for scanning.