Fix log stream cleanup for containers with tty.
When services started with "tty: true" exited early, dcmon was exiting
with the following node error:
    Error [ERR_STREAM_WRITE_AFTER_END]: write after end

The default pipe stream behavior is to close the stream on error but we
were writing to it after that. So change the behavior so that the
behavior to not end the destination when the source ends.